The number of unrelateds is not reported correctly
@odelaneau thanks for sharing this tool, it's very helpful!
Here is a fix of a minor typo leading to incorrectly reported number of unrelated individuals.
Do you have the machine / environment used to build this tool? I'm using the pre-compiled binary  (in the ``bin``) folder, that would be great to get an updated version that include this fix - but if you need to configure build environment again I could also try harder to build it on my end.
Thanks again!